+++ source/B2B.Transactions.IntegrationTests/IncomingMessages/IncomingMessageBuilder.cs
